feat: add VS Code support, Windows/WSL documentation, and Makefile portability
Support developing CNet doors in VS Code on Windows (via WSL) and native
Linux with full IntelliSense, one-key builds, and code snippets.

Makefile portability: both Makefiles now use a configurable PREFIX variable
(defaults to /opt/amiga) so the cross-compiler can be installed anywhere.

VS Code workspace configuration provides IntelliSense by querying the actual
m68k-amigaos-gcc cross-compiler for system include paths and predefined
macros, build tasks with GCC problem matcher for clickable error navigation,
8 code snippets for common CNet door development patterns, and recommended
extensions.

Step-by-step documentation covers the complete journey from WSL installation
through cross-compiler build to deploying compiled doors to WinUAE, with a
parallel guide for native Linux users.

Example files cleaned up to build with zero warnings under -Wall (adds
parentheses around assignments in conditions, removes unused variables, adds
casts for AmigaOS Printf() pointer arguments and format specifier
mismatches).

Closes #1

+++ examples/mainporttest.c
@@ -31,20 +31,20 @@ int main(int argc, char *argv[])
 			account = atoi(argv[1]);
 			Printf("	Examining mail semaphore for account %ld..\n", (long)account);
 			Printf("		Address     : 0x%lx\n", (unsigned long)&myp->MPE->MailSem[account-1]);
-			Printf("		Owner task  : 0x%lx (%s)\n", (unsigned long)myp->MPE->MailSem[account-1].ss_Owner, (taskaddr == myp->MPE->MailSem[account-1].ss_Owner) ? argv[0]:"unknown");
+			Printf("		Owner task  : 0x%lx (%s)\n", (unsigned long)myp->MPE->MailSem[account-1].ss_Owner, (long)((taskaddr == myp->MPE->MailSem[account-1].ss_Owner) ? argv[0]:"unknown"));
 
 			if(AttemptSemaphore(&myp->MPE->MailSem[account-1]))
 				{
 				Printf("	(Semaphore attempable)\n");
-				Printf("		Owner task (temp)  : 0x%lx (%s)\n", (unsigned long)myp->MPE->MailSem[account-1].ss_Owner, (taskaddr == myp->MPE->MailSem[account-1].ss_Owner) ? argv[0]:"unknown");
+				Printf("		Owner task (temp)  : 0x%lx (%s)\n", (unsigned long)myp->MPE->MailSem[account-1].ss_Owner, (long)((taskaddr == myp->MPE->MailSem[account-1].ss_Owner) ? argv[0]:"unknown"));
 				ReleaseSemaphore(&myp->MPE->MailSem[account-1]);
 
 				Printf("	Obtaining Semaphore..\n");
 				ObtainSemaphore(&myp->MPE->MailSem[account-1]);
 				Printf("	(Semaphore obtainable)\n");
-				Printf("		Owner task (temp)  : 0x%lx (%s)\n", (unsigned long)myp->MPE->MailSem[account-1].ss_Owner, (taskaddr == myp->MPE->MailSem[account-1].ss_Owner) ? argv[0]:"unknown");
+				Printf("		Owner task (temp)  : 0x%lx (%s)\n", (unsigned long)myp->MPE->MailSem[account-1].ss_Owner, (long)((taskaddr == myp->MPE->MailSem[account-1].ss_Owner) ? argv[0]:"unknown"));
 				ReleaseSemaphore(&myp->MPE->MailSem[account-1]);
-				Printf("	Owner task (final)  : 0x%lx (%s)\n", (unsigned long)myp->MPE->MailSem[account-1].ss_Owner, (taskaddr == myp->MPE->MailSem[account-1].ss_Owner) ? argv[0]:"unknown");
+				Printf("	Owner task (final)  : 0x%lx (%s)\n", (unsigned long)myp->MPE->MailSem[account-1].ss_Owner, (long)((taskaddr == myp->MPE->MailSem[account-1].ss_Owner) ? argv[0]:"unknown"));
 				}
 			}
 		}
